requirements.txt
altgraph==0.17.4
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
netifaces==0.10.6
packaging==24.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.2.3
wsproto==1.2.0
zipp==3.21.0
